Refactoring, added mutex
